app/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.104.1
# -e git+https://github.com/mayolito-waje/BCS33_ML_NLP_research.git@63cd6ada5fb5cf9a1f0ffbc74724ba221c258ae7#egg=freedom_wall_posts_classifier&subdirectory=app
h11==0.14.0
idna==3.6
joblib==1.3.2
nltk==3.8.1
numpy==1.26.2
pandas==2.1.3
pydantic==2.5.2
pydantic_core==2.14.5
pyenchant==3.2.2
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.10.3
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.24.0.post1
